rename metrics subpackage
