+++ cwidgets/main.go
@@ -2,12 +2,12 @@ package cwidgets
 
 import (
 	"github.com/bcicen/ctop/logging"
-	"github.com/bcicen/ctop/metrics"
+	"github.com/bcicen/ctop/models"
 )
 
 var log = logging.Init()
 
 type WidgetUpdater interface {
 	SetMeta(string, string)
-	SetMetrics(metrics.Metrics)
+	SetMetrics(models.Metrics)
 }
